# Labels computation LogQLv2