--- a/pkg/logql/ast.go
+++ b/pkg/logql/ast.go
@@ -85,7 +85,7 @@ type Pipeline = log.Pipeline
 type SampleExtractor = log.SampleExtractor
 
 var (
-	NoopPipeline = log.NoopPipeline
+	NoopPipeline = log.NewNoopPipeline()
 )
 
 // PipelineExpr is an expression defining a log pipeline.
@@ -109,7 +109,7 @@ func (m MultiStageExpr) Pipeline() (log.Pipeline, error) {
 		return nil, err
 	}
 	if len(stages) == 0 {
-		return log.NoopPipeline, nil
+		return NoopPipeline, nil
 	}
 	return log.NewPipeline(stages), nil
 }
@@ -169,7 +169,7 @@ func (e *matchersExpr) String() string {
 }
 
 func (e *matchersExpr) Pipeline() (log.Pipeline, error) {
-	return log.NoopPipeline, nil
+	return NoopPipeline, nil
 }
 
 func (e *matchersExpr) HasFilter() bool {
@@ -719,7 +719,7 @@ func (e *vectorAggregationExpr) Extractor() (log.SampleExtractor, error) {
 	// inject in the range vector extractor the outer groups to improve performance.
 	// This is only possible if the operation is a sum. Anything else needs all labels.
 	if r, ok := e.left.(*rangeAggregationExpr); ok && e.operation == OpTypeSum {
-		return r.extractor(e.grouping, true)
+		return r.extractor(e.grouping, len(e.grouping.groups) == 0)
 	}
 	return e.left.Extractor()
 }
@@ -860,7 +860,7 @@ func (e *literalExpr) String() string {
 func (e *literalExpr) Selector() LogSelectorExpr               { return e }
 func (e *literalExpr) HasFilter() bool                         { return false }
 func (e *literalExpr) Operations() []string                    { return nil }
-func (e *literalExpr) Pipeline() (log.Pipeline, error)         { return log.NoopPipeline, nil }
+func (e *literalExpr) Pipeline() (log.Pipeline, error)         { return NoopPipeline, nil }
 func (e *literalExpr) Matchers() []*labels.Matcher             { return nil }
 func (e *literalExpr) Extractor() (log.SampleExtractor, error) { return nil, nil }